Corrects a CSS selector.
I haven't tested this change (because my mac isn't letting me download Xcode developer tools).

This change may fix this problem:
The temporary caption window that says "Spanish (autogenerated). Click <> for settings" has the class "caption-window", so the additional captions are attached to that window, which disappears in a second, so the additional captions also disappear.